various bug fixes indentation, missing arguments, invalid syntax made viz install not remove target directory everytime

+++ haros/visualiser.py
@@ -51,20 +51,21 @@ except ImportError:
 _log = logging.getLogger(__name__)
 
 
-def install(dst, source_runner):
-    if os.path.exists(dst):
+def install(dst, source_runner, force = False):
+    if force and os.path.exists(dst):
         rmtree(dst)
     if not os.path.exists(dst):
-        data_dir = os.path.join(dst, "data")
         _log.info("Creating %s", dst)
         os.mkdir(dst)
-        _log.info("Copying viz files.")
-        if source_runner:
-            src = os.path.abspath(os.path.join(os.path.dirname(__file__),
-                                  "..", "harosviz"))
-        else:
-            src = resource_filename(Requirement.parse("haros"), "harosviz")
-        copy_tree(src, dst)
+    _log.info("Copying viz files.")
+    if source_runner:
+        src = os.path.abspath(os.path.join(os.path.dirname(__file__),
+                              "..", "harosviz"))
+    else:
+        src = resource_filename(Requirement.parse("haros"), "harosviz")
+    copy_tree(src, dst)
+    data_dir = os.path.join(dst, "data")
+    if not os.path.exists(data_dir):
         _log.info("Creating %s", data_dir)
         os.mkdir(data_dir)
